duoc_go/requirements.txt
asgiref==3.9.1
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.6
django-tailwind==4.2.0
djangorestframework==3.16.1
idna==3.11
marshmallow==3.26.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
qrcode==8.2
requests==2.32.5
sqlparse==0.5.3
transbank-sdk==6.1.0
typing_extensions==4.15.0
urllib3==2.5.0
